Split MiqRpmPackages based on backing database
