+++ lib/metadata/linux/MiqRpmPackages/Sqlite.rb
@@ -1,0 +1,4 @@
+class MiqRpmPackages
+  class Sqlite < MiqRpmPackages
+  end
+end
